+++ tsconfig.json
@@ -22,6 +22,7 @@
     "paths": {
       "@/*": ["./*"]
     },
+
     "target": "ES2017"
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
